Reglement in die Datenbank (Migration 0008) und Korrektur zu Art. 2.6.001
KORREKTUR
---------
Beide Punkteskalen verweisen auf Artikel 2.6.001 "regarding the duration
of the event". Ich hatte das als Punkteskala nach Renndauer gelesen, also
als fehlende Tabelle. Das war falsch. Kapitel VI liegt jetzt vor, und
2.6.001 ist keine Skala, sondern eine Ausfallregel: bleibt von einem
Etappenrennen nur eine Etappe oder ein Prolog uebrig, zaehlen
ausschliesslich die Etappenpunkte - kein Gesamtklassement, kein Trikot,
keine Nebenwertungen. Es fehlt also keine Tabelle, aber die Berechnung
braucht eine Bedingung, die nur bei abgebrochenen Rennen greift.

WEITERE PUNKTERELEVANTE REGELN AUS KAPITEL VI
---------------------------------------------
Elf Artikel, in docs/uci-punkte.md tabelliert. Die wichtigste:
Art. 2.6.018 verbietet das Fuehrungstrikot am ERSTEN Tag - ein Rennen mit
N Etappen gibt Trikot-Punkte fuer N-1 Etappen, nicht fuer N. Dazu
Art. 2.6.032: einem ausserhalb des Zeitlimits weitergelassenen Fahrer
werden alle Punkte der Nebenwertungen entzogen.

Zwei Fragen bleiben offen und sind als offen benannt, nicht geraten:
welches der bis zu vier Trikots die Punkte gibt (die Tabelle sagt es
nicht, und der Traeger ist nach 2.6.018 nicht immer der Fuehrende), und
ob Mannschaftszeitfahr-Etappen Etappenpunkte geben (2.10.008 sagt ja,
2.6.025 klingt nach nein).

DAS REGLEMENT IN DIE DATENBANK
------------------------------
Die Punkte sollen selbst berechnet werden. Dafuer muessen die Skalen
abfragbar sein - eine CSV genuegt fuer die Dokumentation, nicht fuer ein
JOIN gegen Ergebniszeilen. Migration 0008 legt uci_stufe, uci_punkte,
uci_rennstufe und uci_quelle an; gefuellt werden sie von
app/uci_punkte.py aus der CSV, damit die 1564 Zahlen genau einmal im
Repository stehen.

uci_stufe ist eine eigene Tabelle, weil die beiden anderen per
Fremdschluessel darauf zeigen: eine Rennzuordnung auf eine Stufe ohne
Skala faellt beim Rechnen nicht als Fehler auf, sondern als stilles
Null-Ergebnis.

Der Hash-Merker in uci_quelle ist kein Luxus: ohne ihn wuerden bei jedem
Prozessstart 1564 Zeilen geschrieben, und diese Instanz startet mehrmals
pro Stunde neu (gemessen, Migration 0007).

uci_rennstufe.race_id ist bewusst NULL: der Reglement-Name laesst sich
nicht auf races.id abbilden (das Reglement schreibt "Oomlop Nieuwsblad"
fuer Omloop Nieuwsblad). Die Luecke ist damit abfragbar statt unsichtbar.

rider_season_points.uci_points ist jetzt NUMERIC(8,2):
Mannschaftszeitfahren-Punkte werden auf ein Hundertstel geteilt.
INTEGER -> NUMERIC ist verlustfrei, also auch ohne Backup zulaessig; die
Migration meldet, wie viele Zeilen ueberhaupt einen Wert hatten.

EIN MANGEL IM PRUEF-GERUEST, DEN DAS AUFGEDECKT HAT
---------------------------------------------------
Der Wertevergleich "kein bestehender Spaltenwert veraendert" lag in DREI
Kopien unter drei Namen (check-migrations.sh, -0002.sh, -0003.sh) und
verglich die TEXTDARSTELLUNG der Zeilen, nicht deren Wert. Die
verlustfreie Typerweiterung macht aus "5" ein "5.00" und liess zwei der
drei Kopien scheitern, obwohl sich kein Wert geaendert hatte.

Der Vergleich steht jetzt einmal in scripts/_datenvergleich.sh und
normalisiert Zahlspalten mit trim_scale(...::numeric). Gegengeprueft in
allen drei Skripten mit einer Testmigration, die einen Wert
ueberschreibt: jedes schlaegt weiterhin an.

GEPRUEFT
--------
scripts/check-uci-reglement.py, 21 Pruefungen gegen ein echtes Postgres,
leere Datenbank genuegt: Tabellenform und neuer Spaltentyp, erster
Ladevorgang gegen die Zeilenzahlen der CSV, zweiter Ladevorgang tut
nichts, eine von Hand gesetzte race_id uebersteht einen Reload mit
geaenderter Datei, ein aus der CSV entfernter Wert verschwindet auch aus
der Datenbank, Fremdschluessel und CHECKs greifen, 12,86 Punkte bleiben
12,86.

Gegengeprueft mit drei Manipulationen am Loader - Hash-Merker entfernt,
race_id ueberschrieben, Loeschen veralteter Zeilen ausgebaut - jede
bricht genau die zustaendige Pruefung.

Endpunkt-Regression: 48 Endpunkte, 532 Blattwerte, 0 Unterschiede. Der
Typwechsel ist unsichtbar, weil uci_points ueberall NULL ist. Weiter
gruen: ruff -F, check-vokabular (34 Dateien), check-env-example,
check-uci-punkte, check-migrations und check-migration-0002 bis -0006,
check-kadenz.

WAS FUER DIE BERECHNUNG NOCH FEHLT
----------------------------------
Vollstaendig in docs/uci-punkte.md. Kurz: gc und etappe sind rechenbar,
sobald die 121 race_id-Zuordnungen gesetzt sind. nebenklassement und
trikot sind gar nicht modelliert. Der Kontinentalkalender braucht die
Unterscheidung Class 1 / Class 2 / 1.2U, die taxonomy.Category nicht
kennt. Und sechs der dreizehn Anlaesse haben keine Quelle: nationale
Meisterschaften, Kontinentalmeisterschaften, WM und Olympia stehen nicht
im Kalender.

+++ backend/app/models.py
@@ -92,7 +92,11 @@ class RiderSeason(BaseModel):
     year: int
     team_name: str
     team_wiki_url: Optional[str] = None
-    uci_points: Optional[int] = None
+    # float, nicht int: Mannschaftszeitfahren-Punkte werden auf ein
+    # Hundertstel geteilt (Art. 2.10.008), ein Fahrer kann also 12,86
+    # Punkte haben. Die Spalte ist seit Migration 0008 NUMERIC(8,2);
+    # psycopg liefert Decimal, Pydantic wandelt in float.
+    uci_points: Optional[float] = None
 
 
 class RiderDetail(Rider):
